Restructure stores
