+++ src/components/Home.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import { Link } from 'react-router-dom'
+
+export default function Home () {
+  return (
+    <div><Link to="/form">Form</Link></div>
+  )
+}
